Refactor

+++ contracts/interfaces/IWETH.sol
@@ -4,6 +4,8 @@ pragma solidity >=0.5.0;
 
 interface IWETH {
     function deposit() external payable;
-    function transfer(address to, uint value) external returns (bool);
-    function withdraw(uint) external;
+
+    function transfer(address to, uint256 value) external returns (bool);
+
+    function withdraw(uint256) external;
 }
